+++ crates/kira/src/start_time.rs
@@ -1,6 +1,9 @@
 use std::time::Duration;
 
-use crate::clock::ClockTime;
+use crate::clock::{
+	clock_info::{ClockInfoProvider, WhenToStart},
+	ClockTime,
+};
 
 /// Describes when an action should occur.
 #[derive(Debug, Clone, Copy, PartialEq, Default)]
@@ -14,6 +17,36 @@ pub enum StartTime {
 	/// specific time.
 	ClockTime(ClockTime),
 }
+
+impl StartTime {
+	pub(crate) fn update(
+		&mut self,
+		dt: f64,
+		clock_info_provider: &ClockInfoProvider,
+	) -> WillNeverStart {
+		match self {
+			StartTime::Immediate => {}
+			StartTime::Delayed(time_remaining) => {
+				*time_remaining = time_remaining.saturating_sub(Duration::from_secs_f64(dt));
+				if time_remaining.is_zero() {
+					*self = StartTime::Immediate;
+				}
+			}
+			StartTime::ClockTime(clock_time) => {
+				match clock_info_provider.when_to_start(*clock_time) {
+					WhenToStart::Now => {
+						*self = StartTime::Immediate;
+					}
+					WhenToStart::Later => {}
+					WhenToStart::Never => return true,
+				}
+			}
+		}
+		false
+	}
+}
+
+pub(crate) type WillNeverStart = bool;
 
 impl From<Duration> for StartTime {
 	fn from(v: Duration) -> Self {